Introduce new HOSTDIR, e.g. for python running on the build machine
Sometimes it is necessary to compile software needed by other packages
but running on the host machine, e.g. python. So we need a directory
to install those software in. It will not be copied to the target. Add the
bin directory to PATH and also set PYTHONPATH to the lib directory of
HOSTDIR as Python is one of the first packages which needs this.